requirements.txt
# Common Python dependencies for ADS-B + AMeDAS Lab
# This file is intended for local development on app_server and ADS-B hosts.

# Core database and HTTP client
psycopg2-binary>=2.9,<3.0
requests>=2.31,<3.0
python-dateutil>=2.8,<3.0

# Django-based web viewer (adsb_viewer)
Django>=4.2,<5.0
